refactor: GameResult 로직을 모두 수익금 계산 로직으로 변경

+++ src/main/java/blackjack/domain/participant/Player.java
@@ -37,7 +37,7 @@ public class Player extends Participant {
         stopDrawing = true;
     }
     
-    public long calculateProfit(ProfitRate profitRate) {
+    public int calculateProfit(ProfitRate profitRate) {
         return profitRate.getProfit(bettingAmount);
     }
 }
